# Port RemoveDiagonalGatesBeforeMeasure to rust 