--- a/crates/accelerate/src/lib.rs
+++ b/crates/accelerate/src/lib.rs
@@ -23,6 +23,7 @@ pub mod isometry;
 pub mod nlayout;
 pub mod optimize_1q_gates;
 pub mod pauli_exp_val;
+pub mod remove_diagonal_gates_before_measure;
 pub mod results;
 pub mod sabre;
 pub mod sampled_exp_val;