--- a/crates/accelerate/src/remove_diagonal_gates_before_measure.rs
+++ b/crates/accelerate/src/remove_diagonal_gates_before_measure.rs
@@ -11,8 +11,8 @@
 // that they have been altered from the originals.
 
 /// Remove diagonal gates (including diagonal 2Q gates) before a measurement.
-use pyo3::prelude::*;
 use hashbrown::HashSet;
+use pyo3::prelude::*;
 
 use qiskit_circuit::dag_circuit::{DAGCircuit, NodeType};
 use qiskit_circuit::operations::Operation;