--- a/qiskit/transpiler/passes/optimization/remove_diagonal_gates_before_measure.py
+++ b/qiskit/transpiler/passes/optimization/remove_diagonal_gates_before_measure.py
@@ -38,3 +38,4 @@ class RemoveDiagonalGatesBeforeMeasure(TransformationPass):
             DAGCircuit: the optimized DAG.
         """
         remove_diagonal_gates_before_measure(dag)
+        return dag