@ShellyGarion 2024-09-02
the failing test should be fixed after https://github.com/Qiskit/qiskit/pull/13067 is merged