@ShellyGarion 2024-09-02
Performance comparison of the old python code and the new rust code for a `random_circuit`:

| n_qubits | depth | time is sec of old Python code | time in sec of new rust code | 
|----------|--------|---------------------------|-------------------------|
| 5 | 1024 | 0.06 | 0.00 | 
| 14 | 1024 |  0.47 | 0.02 | 
| 20 | 1024 | 0.92 | 0.04 |